Cleaning project structure
