+++ src/engine/backend/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "kime-engine-backend"
+version = "1.0.0"
+authors = ["Riey <creeper844@gmail.com>"]
+edition = "2018"
+license = "GPL-3.0"
+
+[dependencies]
+ahash = { version = "0.7.0", features = ["serde"] }
+bitflags = "1.2.1"
+enum-map = "0.6.4"
+enumset = { version = "1.0.4", features = ["serde"] }
+serde = { version = "1.0.118", features = ["derive"] }
+serde_yaml = "0.8.15"
+strum = { version = "0.20.0", features = ["derive"] }
