+++ src/engine/backends/hangul/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "kime-engine-hangul"
+name = "kime-engine-backend-hangul"
 version = "0.1.0"
 authors = ["Riey <creeper844@gmail.com>"]
 edition = "2018"
 license = "GPL-3.0"
 
 [dependencies]
-kime-engine-core = { path = "../../core" }
+kime-engine-backend = { path = "../../backend" }
 enumset = "1.0.6"
 num-derive = "0.3.3"
 num-traits = "0.2.14"
