+++ src/engine/backends/hangul/src/layout.rs
@@ -1,6 +1,6 @@
 use crate::characters::KeyValue;
 use crate::Key;
-use kime_engine_core::AHashMap;
+use kime_engine_backend::AHashMap;
 
 #[derive(Clone, Default)]
 pub struct Layout {
