+++ src/engine/backends/latin/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "kime-engine-latin"
+name = "kime-engine-backend-latin"
 version = "0.1.0"
 authors = ["Riey <creeper844@gmail.com>"]
 edition = "2018"
 license = "GPL-3.0"
 
 [dependencies]
-kime-engine-core = { path = "../../core" }
+kime-engine-backend = { path = "../../backend" }
 serde = { version = "1.0.124", features = ["derive"] }
 serde_yaml = "0.8.17"
