+++ src/engine/capi/cbindgen-cpp.toml
@@ -17,7 +17,7 @@ bitflags = true
 
 [parse]
 parse_deps = true
-include = ["kime-engine", "kime-engine-core"]
+include = ["kime-engine-backend", "kime-engine-core"]
 
 [const]
 allow_static_const = true
