--- src/engine/kime-engine/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "kime-engine"
-version = "0.1.0"
-authors = ["Riey <creeper844@gmail.com>"]
-edition = "2018"
-license = "GPL-3.0"
-
-[dependencies]
-kime-engine-hangul = { path = "../backends/hangul" }
-kime-engine-latin = { path = "../backends/latin" }
-kime-engine-dict = { path = "../dict" }
-kime-engine-core = { path = "../core" }
-serde = { version = "1.0.124", features = ["derive"] }
-serde_yaml = "0.8.17"
-enum-map = "0.6.4"
-enumset = "1.0.6"
-
-[target.'cfg(unix)'.dependencies]
-xdg = "2.2.0"
-kime-run-dir = { path = "../../tools/run_dir" }
