+++ src/tools/check/Cargo.toml
@@ -8,7 +8,7 @@ license = "GPL-3.0"
 [dependencies]
 ansi_term = "0.12.1"
 kime-engine-cffi = { path = "../../engine/cffi" }
-kime-engine = { path = "../../engine/kime-engine" }
+kime-engine-core = { path = "../../engine/core" }
 pad = "0.1.6"
 serde_yaml = "0.8.17"
 strum = { version = "0.20.0", features = ["derive"] }
